+++ utopia-remix/package.json
@@ -56,6 +56,7 @@
     "tiny-invariant": "1.3.1",
     "unzipper": "0.11.5",
     "url-join": "5.0.0",
+    "utopia-shared": "link:../utopia-shared",
     "zustand": "4.5.0"
   },
   "devDependencies": {
